botnet/requirements.txt
ping3
pypiwin32
winshell
pytest-shutil
pathlib
pycryptodome
pyinstaller
